+++ ts/features/bonus/bpd/analytics/index.ts
@@ -24,8 +24,6 @@ import {
   bpdDetailsLoadAll,
   bpdLoadActivationStatus
 } from "../store/actions/details";
-import { bpdAmountLoad } from "../store/actions/amount";
-import { bpdPeriodsLoad } from "../store/actions/periods";
 import { bpdSelectPeriod } from "../store/actions/selectedPeriod";
 import {
   bpdPaymentMethodActivation,
@@ -78,14 +76,8 @@ const trackAction = (mp: NonNullable<typeof mixpanel>) => (
         trxDate: action.payload.results.map(r => r.trxDate.toString()),
         circuitType: action.payload.results.map(r => r.circuitType)
       });
-    case getType(bpdTransactionsLoad.failure):
-    case getType(bpdAmountLoad.failure):
-      return mp.track(action.type, {
-        awardPeriodId: action.payload.awardPeriodId,
-        reason: action.payload.error.message
-      });
-
     // CashBack details
+    case getType(bpdTransactionsLoad.failure):
     case getType(bpdDetailsLoadAll):
     case getType(bpdLoadActivationStatus.request):
       return mp.track(action.type);
@@ -97,23 +89,10 @@ const trackAction = (mp: NonNullable<typeof mixpanel>) => (
       return mp.track(action.type, { reason: action.payload.message });
 
     // Amount
-    case getType(bpdAmountLoad.request):
-      return mp.track(action.type, { awardPeriodId: action.payload });
-    case getType(bpdAmountLoad.success):
     case getType(bpdSelectPeriod): // SelectedPeriod
       return mp.track(action.type, {
         awardPeriodId: action.payload.awardPeriodId
       });
-
-    // Period
-    case getType(bpdPeriodsLoad.request):
-      return mp.track(action.type);
-    case getType(bpdPeriodsLoad.success):
-      return mp.track(action.type, {
-        count: action.payload.length
-      });
-    case getType(bpdPeriodsLoad.failure):
-      return mp.track(action.type, { reason: action.payload.message });
 
     // PaymentMethod
     case getType(bpdPaymentMethodActivation.request):